requirements.txt
fastapi
uvicorn
langgraph
sentence-transformers
faiss-cpu
transformers
pydantic
requests